+++ api-derive/balances/types.ts
@@ -1,9 +1,9 @@
 // Copyright 2017-2022 @polkadot/api-derive authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { AccountId, Balance, BalanceLockTo212, Index } from 'https://deno.land/x/polkadot@0.0.4-3/types/interfaces/index.ts';
-import type { PalletBalancesBalanceLock, PalletBalancesReserveData } from 'https://deno.land/x/polkadot@0.0.4-3/types/lookup.ts';
-import type { BN } from 'https://deno.land/x/polkadot@0.0.4-3/util/mod.ts';
+import type { AccountId, Balance, BalanceLockTo212, Index } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { PalletBalancesBalanceLock, PalletBalancesReserveData } from 'https://deno.land/x/polkadot/types/lookup.ts';
+import type { BN } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 export interface DeriveBalancesAccountData {
   freeBalance: Balance;