+++ types-codec/native/Bool.ts
@@ -1,10 +1,10 @@
 // Copyright 2017-2022 @polkadot/types-codec authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.0.4-3/util/types.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 import type { AnyBool, Codec, Inspect, IU8a, Registry } from '../types/index.ts';
 
-import { isU8a, u8aToHex } from 'https://deno.land/x/polkadot@0.0.4-3/util/mod.ts';
+import { isU8a, u8aToHex } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 /**
  * @name bool